chore(sentinel-api-service): apply review feedback
- Rename `getRelayStatus` -> `getSmartTransaction` and `Raw*RelayStatus*Struct` -> `Raw*SmartTransaction*Struct` to match the `/smart-transactions/{uuid}` endpoint path (singular).
- Merge `response.types.ts` into `types.ts`; place service types at the top of `types.ts`. Move messenger and constructor option types into `types.ts`.
- Capitalize the first word of every error message after the `Sentinel API:` prefix.
- Move the sentinel-api-service line under the existing shared-ownership CODEOWNERS section.
- Add endpoint-first debug logs (`endpointName`, `'Request'|'Response'`) to all four endpoints; drop the duplicate internal `#jsonRpc` logs.
