+++ packages/sentinel-api-service/CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add `SentinelApiService` data service centralising all interactions with the MetaMask Sentinel API
   - Simulate transactions via `infura_simulateTransactions` (`simulateTransactions`)
   - Submit gas station relay transactions via `eth_sendRelayTransaction` (`submitRelayTransaction`)
-  - Poll relay transaction status by UUID (`getRelayStatus`)
+  - Look up submitted smart transactions by UUID (`getSmartTransaction`)
   - Fetch and cache the supported-network registry (`getNetworks`)
   - Authenticate requests with client identity headers (`X-Client-Id`, `X-Client-Version`) and a best-effort `Authorization` bearer token from `AuthenticationController:getBearerToken`
   - Target a selectable Sentinel environment via the `environment` constructor option and the `SentinelEnvironment` enum (`Dev`, `Uat`, `Prod`), defaulting to `Prod`
