+++ packages/sentinel-api-service/src/errors.ts
@@ -22,8 +22,8 @@ export class SentinelChainNotSupportedError extends Error {
   constructor(chainId: string, capability?: string) {
     super(
       capability
-        ? `Sentinel API: does not support '${capability}' for chain ${chainId}`
-        : `Sentinel API: does not support chain ${chainId}`,
+        ? `Sentinel API: Does not support '${capability}' for chain ${chainId}`
+        : `Sentinel API: Does not support chain ${chainId}`,
     );
     this.name = 'SentinelChainNotSupportedError';
   }
@@ -37,7 +37,7 @@ export class SentinelChainNotSupportedError extends Error {
  */
 export class SentinelApiResponseValidationError extends Error {
   constructor(message?: string) {
-    super(message ?? 'Sentinel API: malformed response received');
+    super(message ?? 'Sentinel API: Malformed response received');
     this.name = 'SentinelApiResponseValidationError';
   }
 }
